Inventory Slot Update 2
Inventory items now take up a slot and items can only be added to the Inventory if there is atleast one slot open. Items added to a full inventory will be put into the room's chest. Fixed an issue where the itemSubType did not carry over when getting a random loot item.