Code Cleanup
Separated quests and quest functions into main_quests.js. Added a debug button that makes it easier to test the game without having to wait for th e room cooldown. Cleaned up updateArrayItems and added a function for getting an array of a random amount of currency between a min and max for loot. Added a few new locations.