Buy & Sell Update
Added buy and sell multipliers to items when in a location with a shop. Fixed an issue where negative values where not becoming shortened when going through getSmallNumberString. Fixed an issue where the tr_c was scrolling when overlay was enabled. Added a sell all button for selling stacked items at a shop.